Working safe editing and deleting class

+++ src/main/java/pmgkn/pescores/pescores/web/ClassesController.java
@@ -1,6 +1,7 @@
 package pmgkn.pescores.pescores.web;
 
 import jakarta.validation.Valid;
+import org.hibernate.ObjectNotFoundException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -9,7 +10,10 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import pmgkn.pescores.pescores.domain.dto.binding.ClassAddBindingDto;
 import pmgkn.pescores.pescores.domain.dto.binding.ClassEditBindingDto;
 import pmgkn.pescores.pescores.domain.dto.binding.StudentUpdateDto;
+import pmgkn.pescores.pescores.domain.entity.ClassEntity;
+import pmgkn.pescores.pescores.domain.entity.UserEntity;
 import pmgkn.pescores.pescores.service.ClassesService;
+import pmgkn.pescores.pescores.service.UserService;
 
 import java.security.Principal;
 import java.util.UUID;
@@ -20,8 +24,12 @@ public class ClassesController {
 
     private final ClassesService classesService;
 
-    public ClassesController(ClassesService classesService) {
+    private final UserService userService;
+
+    public ClassesController(ClassesService classesService,
+                             UserService userService) {
         this.classesService = classesService;
+        this.userService = userService;
     }
 
     @ModelAttribute("classAddDto")
@@ -36,7 +44,7 @@ public class ClassesController {
 
     @GetMapping
     public String getClasses(Model model,
-                             Principal principal){
+                             Principal principal) {
 
         model.addAttribute("classes", this.classesService.getAllClassesByUser(principal.getName()));
 
@@ -44,15 +52,15 @@ public class ClassesController {
     }
 
     @GetMapping("/add")
-    public String getAddClass(){
+    public String getAddClass() {
         return "add-class";
     }
 
     @PostMapping("/add")
     public String postAddClass(@Valid ClassAddBindingDto classAddBindingDto,
-                                 BindingResult bindingResult,
-                                 RedirectAttributes redirectAttributes,
-                                 Principal principal){
+                               BindingResult bindingResult,
+                               RedirectAttributes redirectAttributes,
+                               Principal principal) {
 
 
         if (bindingResult.hasErrors()) {
@@ -66,38 +74,53 @@ public class ClassesController {
         }
 
         UUID classId = this.classesService.saveClass(classAddBindingDto, principal.getName());
+        UserEntity currentUser = this.userService.getUserByEmail(principal.getName());
 
-        return "redirect:/classes/"+ classId;
+        return "redirect:/classes/" + currentUser.getId() + "/" + classId;
     }
 
-    @GetMapping("/{id}")
-    public String getCurrentClass(@PathVariable("id") UUID id,
+    @GetMapping("/{idTeacher}/{id}")
+    public String getCurrentClass(@PathVariable("idTeacher") UUID idTeacher,
+                                  @PathVariable("id") UUID id,
                                   Model model,
                                   Principal principal) {
 
+        if (!this.userService.getUserById(idTeacher).equals(this.userService.getUserByEmail(principal.getName()))) {
+            throw new ObjectNotFoundException("the info is not yours", ClassEntity.class);
+        }
+
+
         model.addAttribute("classes", this.classesService.getAllClassesByUser(principal.getName()));
         model.addAttribute("currentClass", this.classesService.getClassById(id));
-        model.addAttribute("students",this.classesService.getStudentsSorted(id));
-        model.addAttribute("studentUpdate",new StudentUpdateDto());
+        model.addAttribute("students", this.classesService.getStudentsSorted(id));
+        model.addAttribute("studentUpdate", new StudentUpdateDto());
         model.addAttribute("id", id);
 
         return "current-class";
     }
 
-    @GetMapping("/edit/{id}")
-    public String getEditClass(@PathVariable("id") UUID id,
-                               Model model){
+    @GetMapping("/edit/{idTeacher}/{id}")
+    public String getEditClass(@PathVariable("idTeacher") UUID idTeacher,
+                               @PathVariable("id") UUID id,
+                               Model model,
+                               Principal principal) {
 
-        model.addAttribute("class",this.classesService.getClassById(id));
+        if (!this.userService.getUserById(idTeacher).equals(this.userService.getUserByEmail(principal.getName()))) {
+            throw new ObjectNotFoundException("the info is not yours", ClassEntity.class);
+        }
+
+        model.addAttribute("class", this.classesService.getClassById(id));
 
         return "classes-edit";
     }
 
     @PatchMapping("/edited/{id}")
-    public String editClass(@PathVariable("id") UUID id,
+    public String editClass(
+                            @PathVariable("id") UUID id,
                             @Valid ClassAddBindingDto classEdit,
                             BindingResult bindingResult,
-                            RedirectAttributes redirectAttributes){
+                            RedirectAttributes redirectAttributes,
+                            Principal principal) {
 
         if (bindingResult.hasErrors()) {
 
@@ -105,17 +128,25 @@ public class ClassesController {
             redirectAttributes.addFlashAttribute("org.springframework.validation.BindingResult.classEdit"
                     , bindingResult);
 
-            return "redirect:/classes/edit/{id}";
+            return "redirect:/classes/edit/{idTeacher}/{id}";
 
         }
 
         this.classesService.editClass(id, classEdit);
 
-        return "redirect:/classes/" + id;
+        UserEntity currentUser = this.userService.getUserByEmail(principal.getName());
+
+        return "redirect:/classes/" + currentUser.getId() + "/" + id;
     }
 
-    @DeleteMapping("/delete/{id}")
-    public String deleteClass(@PathVariable("id") UUID id) {
+    @PostMapping("/delete/{idTeacher}/{id}")
+    public String deleteClass(@PathVariable("idTeacher") UUID idTeacher,
+                              @PathVariable("id") UUID id,
+                              Principal principal) {
+
+        if (!this.userService.getUserById(idTeacher).equals(this.userService.getUserByEmail(principal.getName()))) {
+            throw new ObjectNotFoundException("the info is not yours", ClassEntity.class);
+        }
 
         this.classesService.deleteClass(id);
